+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Version](http://semver.org/spec/v2.0.0.htm
 ## [Unreleased]
 ### Added
 ### Changed
+* Bugfix: Update auth.json generation file
 ### Removed
 
 ## [1.1.3] - 25-02-2022
@@ -25,7 +26,7 @@ and this project adheres to [Semantic Version](http://semver.org/spec/v2.0.0.htm
 ## [1.1.1] - 03-10-2021
 ### Added
 * Bugfix: Elasticsearch magento cli installation parameters not available for Magento 2.3.X
-* Bufgix: Unable to connect to mailhog service from Magento  
+* Bugfix: Unable to connect to mailhog service from Magento  
 * Feature: SampleData optionally installable during the project creation 
 * Added RabbitMQ service optionally installable during the project creation