+++ CHANGELOG.md
@@ -6,11 +6,15 @@ and this project adheres to [Semantic Version](http://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 ### Added
+### Changed
+### Removed
+
+## [1.1.4] - 05-05-2022
+### Added
 * Added Support for ElasticSearch 7.16
 ### Changed
 * BugFix: Fix services information displayed after installing Magento, and command bin/describe
 * Bugfix: Update auth.json generation file
-### Removed
 
 ## [1.1.3] - 25-02-2022
 ### Added